Basic message and user api complete
still need auth

+++ databaseOperations/userManipulation.go
@@ -8,20 +8,57 @@ import (
 	"strconv"
 )
 
+// UserExistsId checks if a user exists with a provided id
+func UserExistsId(id int, db *sql.DB) bool {
+	rows, err := db.Query("SELECT id FROM users WHERE id = ?", id)
+
+	if err != nil {
+		log.Fatalf("Error: %s", err)
+	}
+
+	if rows.Next() {
+		return true
+	}
+
+	return false
+}
+
+// UserExistsEmail checks if a user exists with a provided email
+func UserExistsEmail(email string, db *sql.DB) bool {
+	rows, err := db.Query("SELECT email FROM users WHERE email = ?", email)
+
+	if err != nil {
+		log.Fatalf("Error: %s", err)
+	}
+
+	if rows.Next() {
+		return true
+	}
+
+	return false
+}
+
 func CreateUser(user types.User, db *sql.DB) (int, error) {
 	rows, _ := db.Query("SELECT email from users WHERE email = ?", user.Email)
 	if !rows.Next() {
 		_, err := db.Exec("INSERT INTO users (first_name, last_name, email, password) VALUES(?,?,?,?)", user.FirstName, user.LastName, user.Email, user.Password)
+
 		if err != nil {
 			log.Fatalf("Error: %s", err)
 		}
+
 		rows, err = db.Query("SELECT id FROM users WHERE email = ?", user.Email)
+
+		if err != nil {
+			log.Fatalf("Error: %s", err)
+		}
 
 		var id int
 		rows.Next()
 		err = rows.Scan(&id)
+
 		if err != nil {
-			panic(err.Error())
+			log.Fatalf("Error: %s", err)
 		}
 
 		return id, nil
@@ -50,36 +87,37 @@ func GetUser(email string, db *sql.DB) (types.User, error) {
 func EditUser(user types.User, db *sql.DB) error {
 	// User with the new email already exists and is not itself
 	rows, err := db.Query("SELECT email from users WHERE email = ? AND id != ?", user.Email, user.Id)
+
+	if err != nil {
+		log.Fatalf("Error: %s", err)
+		return err
+	}
+
 	if rows.Next() != true {
 		// User does not change its email to a conflicting one or is not changing it
-		rows, err = db.Query("SELECT id FROM users WHERE id = ?", user.Id)
-
-		if err != nil {
-			log.Fatalf("Error: %s", err)
-			return err
-		}
-
-		if rows.Next() {
+		if UserExistsId(user.Id, db) {
 			_, err = db.Exec("UPDATE users SET first_name = ?, last_name = ?, email = ?, password = ? WHERE id = ?", user.FirstName, user.LastName, user.Email, user.Password, user.Id)
 			if err != nil {
 				log.Fatalf("Error: %s", err)
 				return err
 			}
-			return nil
-		} else {
-			return errors.New("user with this ID does not exist")
 		}
 	} else {
 		return errors.New("user with this email already exists")
 	}
+	return nil
 }
 
 // DeleteUser Deletes user in db argument based on provided email argument
 func DeleteUser(email string, db *sql.DB) error {
-	_, err := db.Exec("DELETE FROM users WHERE email = ?", email)
-	if err != nil {
-		log.Fatalf("Error: %s", err)
-		return err
+	if UserExistsEmail(email, db) {
+		_, err := db.Exec("DELETE FROM users WHERE email = ?", email)
+		if err != nil {
+			log.Fatalf("Error: %s", err)
+			return err
+		}
+	} else {
+		return errors.New("user does not exist")
 	}
 	return nil
 }
